Use a blueprint for error handling
