+++ app/__init__.py
@@ -27,9 +27,7 @@ def create_app(config_class=Config):
     from app.api import bp as api_bp
     app.register_blueprint(api_bp, url_prefix='/api/v1')
 
-    from app.api.errors import page_not_found
-    from app.api.errors import internal_server_error
-    app.register_error_handler(404, page_not_found)
-    app.register_error_handler(500, internal_server_error)
+    from app.errors import bp as error_bp
+    app.register_blueprint(error_bp)
 
     return app
