+++ app/errors/handlers.py
@@ -1,7 +1,10 @@
 from flask import jsonify
+from app.errors import bp
+from app import db
 
 
 # Error Handlers
+@bp.app_errorhandler(404)
 def page_not_found(e):
     error = {
         "errors": [
@@ -15,6 +18,7 @@ def page_not_found(e):
     return jsonify(error), 404
 
 
+@bp.app_errorhandler(500)
 def internal_server_error(e):
     error = {
         "errors": [
@@ -26,3 +30,11 @@ def internal_server_error(e):
     }
     # Set the 500 status explicitly
     return jsonify(error), 500
+
+
+@bp.teardown_request
+def teardown_request(exception=None):
+    if exception:
+        print('exception', exception)
+        db.session.rollback()
+    db.session.remove()
